tux3: Remove PTE modification for forked page
With copy-on-write aware ->page_mkwrite(), we don't need to update PFN
of PTE on faulted address (caller of ->page_mkwrite() takes care of it).

So, this removes the modification of PTE on faulted address from
current ->page_mkwrite() related code. (I.e. pass vma of faulted
address to page_cow_file(), and if vma is for faulted address, it
doesn't update PFN in that vma.)

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>