Add handler for page truncation on truncate_inode_pages_range()
truncate_inode_pages_range() is assuming, FS has to wait the
PageWriteback(), but it is not true if page is handled by page forking
at least. (Also, data=journal may be able to do some if want.)  And
there are some unexported functions in truncate_inode_pages_range(),
so FS can't replace it with per-FS code.

So, this adds handler (->truncatepage) to allow implement per-FS way
to truncate and release page.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>