tux3: Make iowait more flexible, then usable in commit.c too
Now, iowait structure is used in buffer_writeback.c. But this
structure is useful to pass backend information.

Now, we want to use this to pass "flush_flags". So rename from iowait
to ioinfo, then save flush_flags. Then separate from
buffer_writeback.c (with this, we can share this code with userland).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>