tux3: ->present removal
To use ->present in frontend, we have to track attributes changes. It
is hard to do, and fragile.

So, this removes ->present. Instead, we check current attributes at
saving into ileaf. And conditionally save attributes.

For example, we store btree attribute only if inode is having dtree.

With this change, we would be able to reduce average attributes size.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>