tux3: Handle ENOSPC for orphaned inode
Unlink/rmdir needs to think about cost of orphaned inode additionally,
not only dirent removal.

	1) dirent removal
	2) inode deletion
	3) adding entry to otree
	4) removing entry from otree

But orphaned inode can be pinned by user control. So, we have to keep
cost for orphan until finishing real deletion of inode. To keep cost,
we track number of orphans, and calculate orphan_cost from it. Then,
we think the orphan_cost as non-free blocks.  (because orphan is user
control, have to return ENOSPC, instead of waiting).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>