Add page forking page_pagefork_file() helper for ->page_mkwrite() handler
This copies helpers from tux3.

When do page forking, we have to update PFN in PTEs of target page. To
do it, this adds page_pagefork_file() helper (and page_pagefork_one()).

page_pagefork_file() takes old and new pages, and updates PFN on related
PTEs of oldpage to newpage. If orig_vma was passed, it excludes to
update PFN (caller of ->page_mkwrite() takes care of it).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>